requirements.txt
# Core dependencies
streamlit>=1.32.0
langchain>=0.1.12
langchain-google-genai>=0.0.11
langchain-openai>=0.0.8
langchain-community>=0.0.27
langchain-qdrant>=0.0.1

# LLM and Embeddings
google-generativeai>=0.3.2
openai>=1.12.0

# Vector Store
qdrant-client>=1.7.3

# Document Processing
pypdf>=4.1.0
python-dotenv>=1.0.1

# Utilities
pyyaml>=6.0.1
python-dateutil>=2.8.2
typing-extensions>=4.10.0

# Development
black>=24.2.0
pylint>=3.0.3
pytest>=8.0.2
